fix: int type overflows

+++ src/concurrent_ht.c
@@ -190,8 +190,8 @@ int HTSize(struct HTNode** hash_table, int size) {
     return counter;
 }
 
-long long int HTProductIDSum(struct HTNode** hash_table, int size) {
-    long long int sum = 0;
+long int HTProductIDSum(struct HTNode** hash_table, int size) {
+    long int sum = 0;
     for(int i = 0; i < size; i++) {
         if(hash_table[i]->productID != -1) sum+=hash_table[i]->productID;
     }
